requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
PyMySQL==1.1.1
cryptography==41.0.7
pycryptodome==3.19.0
Pillow==10.4.0
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.1
hypothesis==6.92.1
coverage==7.3.2